+++ src/trx/game/game_flow/util.c
@@ -54,3 +54,8 @@ void GF_DisableObjectsIfNeeded(void)
         }
     }
 }
+
+int32_t GF_BadGetLevelNum(void)
+{
+    return GF_GetCurrentLevel()->num;
+}