+++ src/trx/game/game_flow/util.h
@@ -1,3 +1,6 @@
 #pragma once
 
 void GF_DisableObjectsIfNeeded(void);
+
+// NOTE: This API should be eliminated!
+int32_t GF_BadGetLevelNum(void);