+++ src/trx/game/objects/effects/flame.h
@@ -1,0 +1,8 @@
+#pragma once
+
+typedef enum {
+    FLAME_BIG,
+    FLAME_SMALL,
+    FLAME_JET,
+    FLAME_SIDE,
+} FLAME_TYPE;